small fix to supress a warning, while forwarding mail
git-svn-id: http://svn.egroupware.org/egroupware/trunk@11929 923e3c2e-ae13-0410-ba18-ced7211e33de

+++ felamimail/inc/class.bocompose.inc.php
@@ -524,7 +524,10 @@
 				while(list($key,$value) = @each($this->sessionData['attachments']))
 				{
 					#print "$key: ".$value['file']."<br>";
-					unlink($value['file']);
+					if (!empty($value['file']))	// happens when forwarding mails
+					{
+						unlink($value['file']);
+					}
 				}
 			}
